# internal/repository/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Subscription struct {
	ID          pgtype.UUID
	ServiceName string
	Price       int32
	UserID      pgtype.UUID
	StartDate   string
	EndDate     pgtype.Text
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}